新闻页面添加

+++ dev-tools-core/src/main/java/xb/dev/tools/base/BaseController.java
@@ -5,11 +5,11 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import javax.servlet.http.HttpServletRequest;
 
 /**
- * @Author: Created by huangxb on 2018-08-03 15:59:06
+ * @author Created by huangxb on 2018-08-03 15:59:06
  * @Description: 基础控制类
  */
 public class BaseController {
-    protected ThreadLocal<Integer> localUser = new ThreadLocal<>();
+    protected ThreadLocal<Long> localUser = new ThreadLocal<>();
 
     @ModelAttribute
     public void loginUserInfo(HttpServletRequest request){
@@ -17,7 +17,7 @@ public class BaseController {
         if(userId!=null) {
             boolean b = userId.matches("[0-9]+");
             if (b) {
-                localUser.set(userId == null ? null : Integer.valueOf(userId));
+                localUser.set(userId == null ? null : Long.valueOf(userId));
             }
         }
     }
